move Flow abstraction to cms/.
